[Refactor] Made Homebrew compile
- Removed unused Win64 files for view provider and input
- Added missing libs in the graphics premake
- Added precompile for Engine and Gameplay
- Minor tweaks to make the sln compile

+++ source/graphics/common/baseshader.cpp
@@ -78,7 +78,7 @@ BaseShader::VS_PS_Blobs BaseShader::CompileShaders(ID3D11Device* device, const S
     popAssert(!FAILED(result), "Pixel Shader Compilation Failed.");
     if (errorMessage)
     {
-        u32 size = errorMessage->GetBufferSize();
+        u32 size = static_cast<u32>(errorMessage->GetBufferSize());
         char* errorMsg = new char[size];
         memcpy(errorMsg, errorMessage->GetBufferPointer(), size);
         int b = 0;
